requirements.txt
autopep8==1.6.0
black==22.1.0
boto3==1.21.21
boto3-stubs==1.21.21
botocore==1.24.21
botocore-stubs==1.24.22
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
flake8==4.0.1
idna==3.3
jmespath==1.0.0
mccabe==0.6.1
mypy-boto3-ec2==1.21.20
mypy-boto3-s3==1.21.7
mypy-extensions==0.4.3
numpy==1.22.3
pandas==1.3.5
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
py4j==0.10.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyspark==3.2.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
s3transfer==0.5.2
schedule==1.1.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
